requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=4.0.0
Pillow>=10.0.0
requests>=2.31.0
django-jazzmin>=3.0.0
cloudinary>=1.36.0
django-cloudinary-storage>=0.3.0
python-dotenv>=1.0.0
